# PERF: faster dir() calls