--- a/pandas/core/accessor.py
+++ b/pandas/core/accessor.py
@@ -24,14 +24,7 @@ class DirNamesMixin:
         """
         Add additional __dir__ for this object.
         """
-        rv = set()
-        for accessor in self._accessors:
-            try:
-                getattr(self, accessor)
-                rv.add(accessor)
-            except AttributeError:
-                pass
-        return rv
+        return {accessor for accessor in self._accessors if hasattr(self, accessor)}
 
     def __dir__(self) -> List[str]:
         """