--- a/pandas/core/generic.py
+++ b/pandas/core/generic.py
@@ -5439,11 +5439,7 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         add the string-like attributes from the info_axis.
         If info_axis is a MultiIndex, it's first level values are used.
         """
-        additions = {
-            c
-            for c in self._info_axis.unique(level=0)[:100]
-            if isinstance(c, str) and c.isidentifier()
-        }
+        additions = self._info_axis._dir_additions_for_owner
         return super()._dir_additions().union(additions)
 
     # ----------------------------------------------------------------------