--- a/pandas/core/generic.py
+++ b/pandas/core/generic.py
@@ -5439,8 +5439,10 @@ class NDFrame(PandasObject, SelectionMixin, indexing.IndexingMixin):
         add the string-like attributes from the info_axis.
         If info_axis is a MultiIndex, it's first level values are used.
         """
-        additions = self._info_axis._dir_additions_for_owner
-        return super()._dir_additions().union(additions)
+        additions = super()._dir_additions()
+        if self._info_axis._can_hold_strings:
+            additions &= self._info_axis._dir_additions_for_owner
+        return additions
 
     # ----------------------------------------------------------------------
     # Consolidation of internals