--- a/pandas/core/indexes/category.py
+++ b/pandas/core/indexes/category.py
@@ -162,7 +162,7 @@ class CategoricalIndex(ExtensionIndex, accessor.PandasDelegate):
     _typ = "categoricalindex"
 
     @property
-    def _can_hold_strings(self) -> bool:
+    def _can_hold_strings(self):
         return self.categories._can_hold_strings
 
     codes: np.ndarray