--- a/pandas/core/indexes/interval.py
+++ b/pandas/core/indexes/interval.py
@@ -2,7 +2,7 @@
 from functools import wraps
 from operator import le, lt
 import textwrap
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple, Union, cast
+from typing import TYPE_CHECKING, Any, List, Optional, Set, Tuple, Union, cast
 
 import numpy as np
 
@@ -370,6 +370,10 @@ class IntervalIndex(IntervalMixin, ExtensionIndex):
     @cache_readonly
     def _multiindex(self) -> MultiIndex:
         return MultiIndex.from_arrays([self.left, self.right], names=["left", "right"])
+
+    @cache_readonly
+    def _dir_additions_for_owner(self) -> Set[str]:
+        return set()
 
     @cache_readonly
     def values(self) -> IntervalArray: